// src/controllers/types/informasi-publik.type.ts
export type InformasiPublikProps = {
  content_id: string;
  type_id: string;
  JenisLayanan: null | string;
  title: string;
  template_id: null | string;
  category_id: string;
  content: string;
  urlcontent: string;
  author: string;
  sumber_informasi: string;
  lampiran: null | string;
  uploaddokumen: string;
  jenis_lampiran: null | string;
  status: string;
  slug_title: string;
  organisasi_id: string;
  ParentId: string;
  ref_id: null | string;
  tgl_publish: string;
  created_at: string;
  updated_at: string;
  site_id: string;
  SiteName: string;
  DomainSite: string;
  Kanal: string;
  Category: string;
  Author: string;
  TypeLayanan: null | string;
};

export type InformasiSertaMertaProps = {
  content_id: string;
  type_id: string;
  JenisLayanan: null | string;
  title: string;
  template_id: null | string;
  category_id: string;
  content: string;
  urlcontent: string;
  author: string;
  sumber_informasi: string;
  lampiran: null | string;
  uploaddokumen: null | string;
  jenis_lampiran: null | string;
  status: string;
  slug_title: string;
  organisasi_id: string;
  ParentId: string;
  ref_id: null | string;
  tgl_publish: string;
  created_at: string;
  updated_at: string;
  site_id: string;
  SiteName: string;
  DomainSite: string;
  Kanal: string;
  Category: string;
  Author: string;
  TypeLayanan: null | string;
};

export type InformasiSetiapSaatProps = {
  content_id: string;
  type_id: string;
  JenisLayanan: null | string;
  title: string;
  template_id: null | string;
  category_id: string;
  content: string;
  urlcontent: string;
  author: string;
  sumber_informasi: string;
  lampiran: null | string;
  uploaddokumen: null | string;
  jenis_lampiran: null | string;
  status: string;
  slug_title: string;
  organisasi_id: string;
  ParentId: string;
  ref_id: null | string;
  tgl_publish: string;
  created_at: string;
  updated_at: string;
  site_id: string;
  SiteName: string;
  DomainSite: string;
  Kanal: string;
  Category: string;
  Author: string;
  TypeLayanan: null | string;
};